revert duplicated currency format number changes
